refactor: split instructor dashboard state
